# swagger.yaml
swagger: "2.0"
info:
  description: "Authors Haven - A Social platform for the creative at heart."
  version: "1.0.0"
  title: "Authors Haven"
basePath: /api/v1/
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /articles/rate/{slug}:
    post:
      tags:
        - Articles
      summary: rate an article
      operationId: rateArticle
      produces:
        - application/json
      parameters:
        - name: "slug"
          in: "path"
          description: An article slug to be rated by a user
          required: true
          type: string
        - name: x-access-token
          in: header
          description: Authorization token
          required: true
          type: string
        - name: "rate"
          in: body
          description: A user rating
          required: true
          schema:
            $ref: '#/definitions/ratings'
      responses:
        '201':
          description: Article rated successfully
        '400':
          description: "Invalid rating supplied"
    get:
      tags:
        - Articles
      summary: get all specific article ratings
      operationId: getArticleRatings
      consumes:
        - application/json
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: "slug"
          in: "path"
          description: An article slug to be rated by a user
          required: true
          type: string
      responses:
        '200':
          description: All ratings for Article
        '404':
          description: "Article does not exist"
...
  /articles/vote/{slug}:
    post:
      tags:
        - "Articles"
      summary: Vote an Article
      description: Vote an article
      operationId: voteArticle
      produces:
        - application/json
      parameters:
        - name: slug
          in: path
          description: Slug of article to vote
          required: true
          type: string
        - name: x-access-token
          in: header
          description: Authorization token
          required: true
          type: string
        - name: "status"
          in: body
          description: A user vote status
          required: true
          schema:
            $ref: '#/definitions/voteArticle'
      responses:
        200:
          description: You have unvote this article
        201:
          description: You upvote this article
        400:
          description: Pass in the appropriate status (true or false)
        404:
          description: This article does not exist
...
  /articles/bookmark/{slug}:
    get:
      tags:
        - "Articles"
      summary: Bookmark an Article
      description: Bookmark an article
      operationId: BookmarkArticle
      produces:
        - application/json
      parameters:
        - name: slug
          in: path
          description: Slug of article to bookmark
          required: true
          type: string
        - name: x-access-token
          in: header
          description: Authorization token
          required: true
          type: string
      responses:
        200:
          description: Bookmark successfully removed
        201:
          description: Bookmark created successfully
        400:
          description: Input your content
        404:
          description: Article does not exist
...
definitions:
...
  ratings:
    type: object
    required:
      - rate
    properties:
      rate:
        type: string
    xml:
      name: rateArticle
...
  voteArticle:
    type: object
    required:
      - status
    properties:
      status:
        type: string
    xml:
      name: voteArticle